Improved Post model
I got rid of all queries with expensive joins and replaced them with simple paginated queries.
The Laravel backend now obtains and returns the username of the post owner as an appendable attribute instead of doing expensive SQL joins.
This change improves performance and code readability.
I also made a significant improvement to the factory method for Post. Before I could not create any posts in tests or in PsyShell due to the foreign key for the owner not having any default value but that has been taken care of.